+++ src/viz/GridDrawer.java
@@ -39,7 +39,7 @@ public class GridDrawer {
 			//boolean bReverseGrid = false;
 
 			if (m_dt.m_treeDrawer.m_bRootAtTop) {
-				int nW = m_dt.getWidth();
+				int nW = (int) m_dt.getWidth();
 				if (m_nGridMode == GridMode.SHORT) {
 					nW = 10;
 				}
@@ -105,7 +105,7 @@ public class GridDrawer {
 					}
 				}
 			} else {
-				int nH = m_dt.getHeight();
+				int nH = (int) m_dt.getHeight();
 				if (m_nGridMode == GridMode.SHORT) {
 					nH = 10;
 				}
